requirements.txt
asgiref==3.2.10
Django==3.1
django-admin-rangefilter==0.6.2
pytz==2020.1
sqlparse==0.3.1
